requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.4.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.0
whitenoise==6.6.0
gunicorn==21.2.0
django-admin-interface==0.28.0
colorfield==0.11.0
django-filter==23.5
